# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediQ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/queue/providers/{provider_id}": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Provider Queue",
        "operationId": "get_provider_queue_api_queue_providers__provider_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueSnapshotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/queue/overview": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Queue Overview",
        "operationId": "get_queue_overview_api_queue_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOverviewOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/status/{token_no}": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Public Status",
        "description": "Public, no login required \u2014 this is the token board patients check.",
        "operationId": "get_public_status_api_status__token_no__get",
        "parameters": [
          {
            "name": "token_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/start": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Start Visit",
        "operationId": "start_visit_api_visits__visit_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/delay": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Delay Visit",
        "operationId": "delay_visit_api_visits__visit_id__delay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/complete": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Complete Visit",
        "operationId": "complete_visit_api_visits__visit_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/no-show": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Mark No Show",
        "operationId": "mark_no_show_api_visits__visit_id__no_show_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/priority": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Set Priority",
        "description": "Staff only. See PriorityRequest \u2014 no symptom/clinical field exists here.",
        "operationId": "set_priority_api_visits__visit_id__priority_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sim/clock": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Sim Clock",
        "operationId": "get_sim_clock_api_sim_clock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimClockOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sim/advance": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Advance Sim Clock",
        "operationId": "advance_sim_clock_api_sim_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimClockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sim/freeze": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Freeze Sim Clock",
        "description": "Wraps app.core.clock.freeze \u2014 added for the console's Freeze/Resume\ncontrols; the underlying clock logic is already implemented and unit-tested\n(see backend/tests/test_clock.py), only the HTTP surface is still M8.",
        "operationId": "freeze_sim_clock_api_sim_freeze_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimClockOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sim/resume": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Resume Sim Clock",
        "operationId": "resume_sim_clock_api_sim_resume_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimClockOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sim/reset": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Reset Sim Clock",
        "operationId": "reset_sim_clock_api_sim_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimClockOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sim/seed": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Reseed Demo Data",
        "operationId": "reseed_demo_data_api_sim_seed_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvanceRequest": {
        "properties": {
          "minutes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Minutes"
          }
        },
        "type": "object",
        "required": [
          "minutes"
        ],
        "title": "AdvanceRequest"
      },
...
      "DelayRequest": {
        "properties": {
          "minutes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Minutes"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "minutes",
          "reason"
        ],
        "title": "DelayRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriorityRequest": {
        "properties": {
          "flag": {
            "type": "boolean",
            "title": "Flag"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "flag",
          "reason"
        ],
        "title": "PriorityRequest",
        "description": "Staff-only. There is no symptom or clinical field here or anywhere in the\nAPI \u2014 priority is an explicit operational decision, never inferred."
      },
...
      "ProviderQueueSummaryOut": {
        "properties": {
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "department_code": {
            "type": "string",
            "title": "Department Code"
          },
          "current_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Token"
          },
          "load": {
            "type": "integer",
            "title": "Load"
          },
          "delayed_count": {
            "type": "integer",
            "title": "Delayed Count"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "provider_name",
          "department_code",
          "current_token",
          "load",
          "delayed_count"
        ],
        "title": "ProviderQueueSummaryOut",
        "description": "One row of GET /api/queue/overview \u2014 summary only, no per-visit detail."
      },
...
      "PublicStatusOut": {
        "properties": {
          "token_no": {
            "type": "string",
            "title": "Token No"
          },
          "department_code": {
            "type": "string",
            "title": "Department Code"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "status": {
            "$ref": "#/components/schemas/VisitStatus"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "estimated_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Start"
          },
          "estimated_wait_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Min"
          },
          "eta_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Reason"
          }
        },
        "type": "object",
        "required": [
          "token_no",
          "department_code",
          "provider_name",
          "status",
          "position",
          "estimated_start",
          "estimated_wait_min",
          "eta_reason"
        ],
        "title": "PublicStatusOut",
        "description": "Response for the public, no-login token lookup. Deliberately excludes\npatient_name/phone and any other provider's queue state."
      },
...
      "QueueOverviewOut": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderQueueSummaryOut"
            },
            "type": "array",
            "title": "Providers"
          }
        },
        "type": "object",
        "required": [
          "providers"
        ],
        "title": "QueueOverviewOut"
      },
      "QueueSnapshotOut": {
        "properties": {
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "department_code": {
            "type": "string",
            "title": "Department Code"
          },
          "current_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Token"
          },
          "next_tokens": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Tokens"
          },
          "load": {
            "type": "integer",
            "title": "Load"
          },
          "delayed_count": {
            "type": "integer",
            "title": "Delayed Count"
          },
          "queue": {
            "items": {
              "$ref": "#/components/schemas/QueueVisitOut"
            },
            "type": "array",
            "title": "Queue"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "provider_name",
          "department_code",
          "current_token",
          "next_tokens",
          "load",
          "delayed_count",
          "queue"
        ],
        "title": "QueueSnapshotOut",
        "description": "Full live state of one provider's queue. Returned by\nGET /api/queue/providers/{id} and pushed over SSE on every recompute."
      },
      "QueueVisitOut": {
        "properties": {
          "visit_id": {
            "type": "integer",
            "title": "Visit Id"
          },
          "token_no": {
            "type": "string",
            "title": "Token No"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          },
          "status": {
            "$ref": "#/components/schemas/VisitStatus"
          },
          "source": {
            "$ref": "#/components/schemas/VisitSource"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "estimated_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Start"
          },
          "estimated_wait_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Min"
          },
          "eta_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Reason"
          },
          "priority_flag": {
            "type": "boolean",
            "title": "Priority Flag"
          }
        },
        "type": "object",
        "required": [
          "visit_id",
          "token_no",
          "patient_name",
          "status",
          "source",
          "position",
          "estimated_start",
          "estimated_wait_min",
          "eta_reason",
          "priority_flag"
        ],
        "title": "QueueVisitOut",
        "description": "One entry in a provider's live queue, as returned inside QueueSnapshotOut."
      },
...
      "SimClockOut": {
        "properties": {
          "effective_time": {
            "type": "string",
            "format": "date-time",
            "title": "Effective Time"
          },
          "offset_minutes": {
            "type": "integer",
            "title": "Offset Minutes"
          },
          "is_frozen": {
            "type": "boolean",
            "title": "Is Frozen"
          }
        },
        "type": "object",
        "required": [
          "effective_time",
          "offset_minutes",
          "is_frozen"
        ],
        "title": "SimClockOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisitOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "slot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "source": {
            "$ref": "#/components/schemas/VisitSource"
          },
          "token_no": {
            "type": "string",
            "title": "Token No"
          },
          "status": {
            "$ref": "#/components/schemas/VisitStatus"
          },
          "scheduled_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Start"
          },
          "checked_in_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked In At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "delay_minutes": {
            "type": "integer",
            "title": "Delay Minutes"
          },
          "delay_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delay Reason"
          },
          "priority_flag": {
            "type": "boolean",
            "title": "Priority Flag"
          },
          "priority_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Reason"
          },
          "estimated_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Start"
          },
          "estimated_wait_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Min"
          },
          "eta_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "provider_id",
          "service_id",
          "slot_id",
          "source",
          "token_no",
          "status",
          "scheduled_start",
          "checked_in_at",
          "started_at",
          "completed_at",
          "delay_minutes",
          "delay_reason",
          "priority_flag",
          "priority_reason",
          "estimated_start",
          "estimated_wait_min",
          "eta_reason",
          "created_at"
        ],
        "title": "VisitOut",
        "description": "Shared representation of a visit, returned by booking, walk-in, and lifecycle\nendpoints. Never includes patient contact details beyond what the caller's role\nis entitled to \u2014 the public status endpoint uses PublicStatusOut instead."
      },
      "VisitSource": {
        "type": "string",
        "enum": [
          "appointment",
          "walkin"
        ],
        "title": "VisitSource"
      },
      "VisitStatus": {
        "type": "string",
        "enum": [
          "booked",
          "checked_in",
          "in_service",
          "completed",
          "cancelled",
          "no_show"
        ],
        "title": "VisitStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}